% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{dataGenerator_SPDw}
\alias{dataGenerator_SPDw}
\title{dataGenerator_SPDw}
\usage{
dataGenerator_SPDw(chpts, d = 5, eigenvalues_mat = sample(5))
}
\arguments{
\item{chpts}{a vector of increasing change-point indices (the last value is data length)}

\item{d}{dimension of the SPD matrix (d x d)}

\item{eigenvalues_mat}{matrix of eigenvalues for successive segments, dimension d x nb_seg with nb_seg the number of segments}
}
\value{
a list of SPD matrices of length 'last value in chpts'
}
\description{
Generating time series for change detection in SPD matrices with Wishart distribution.
Each data point is a matrix obtained by the the realisation of a Wishart distribution.
}
\examples{
dataGenerator_SPDw(chpts = c(20,40), d = 5, eigenvalues_mat = matrix(c(sample(5),2*sample(5)),5,2))
}
