% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_SPD}
\alias{dist_SPD}
\title{dist_SPD}
\usage{
dist_SPD(S1, S2)
}
\arguments{
\item{S1}{a SPD matrix}

\item{S2}{a SPD matrix}
}
\value{
the square root of the sum of log(lambda_i)^2 with lambda_i the eigenvalues of "S1^(-1/2) * S2 * S1^(-1/2)"
}
\description{
Computing distances between symmetric positive definite matrices (with invariant properties)
}
\examples{
n <- 5
p <- 10
M1 <- matrix(rnorm(p*n), n, p)
M2 <- matrix(rnorm(p*n), n, p)
S1 <- M1\%*\%t(M1)
S2 <- M2\%*\%t(M2)
dist_SPD(S1,S2)

}
