% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph.cpt.mean}
\alias{graph.cpt.mean}
\alias{graph_cpt_mean}
\title{graph.cpt.mean}
\usage{
graph_cpt_mean(y, A, states, beta = 0)
}
\arguments{
\item{y}{univariate data vector y}

\item{A}{transition matrix (filled with O and 1)}

\item{states}{vector of state means, a fixed mean value for each node}

\item{beta}{penalty value for transition between two different states/nodes}
}
\value{
A list
}
\description{
change point inference for univariate vector constraint by a graph structure and fixed means
}
\examples{
n <- 5
p <- 10

}
