% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph_cpt_manifold}
\alias{graph_cpt_manifold}
\title{graph_cpt_manifold}
\usage{
graph_cpt_manifold(dists, A, beta = 0)
}
\arguments{
\item{dists}{matrix of distances between data and states, size d x n for d states and n data points}

\item{A}{transition matrix (filled with O and 1)}

\item{beta}{penalty value for transition between two different states/nodes}
}
\value{
A list
}
\description{
change point inference for univariate vector constraint by a graph structure and distances computed on a manifold structure
}
