% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{graph_cpt_mean}
\alias{graph_cpt_mean}
\title{graph_cpt_mean}
\usage{
graph_cpt_mean(y, A, states, beta = 0)
}
\arguments{
\item{y}{univariate data vector y}

\item{A}{transition matrix (filled with O and 1)}

\item{states}{vector of state means, a fixed mean value for each node}

\item{beta}{penalty value for transition between two different states/nodes}
}
\value{
A list
}
\description{
change point inference for univariate vector constraint by a graph structure and fixed means
}
\examples{
vec <- c(rnorm(5, sd = 1), rnorm(5, mean = 1, sd = 0), rnorm(5, mean = 1.9, sd = 1))
states <- c(0,1,2)
res <- graph_cpt_mean(y = vec, A = transition_matrix(3),states = c(0,1,2), beta = 0)
res
}
