% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generator.R
\name{SPD_matrix}
\alias{SPD_matrix}
\title{SPD_matrix}
\usage{
SPD_matrix(eigenvalues)
}
\arguments{
\item{eigenvalues}{vector of eigenvalues}
}
\value{
One SPD matrix
}
\description{
Generating a symmetric positive definite (SPD) matrix.
A diagonal matrix with \code{eigenvalues} on its diagonal is generated.
The rotation is performed using an orthogonal matrix obtained by the QR decomposition of a random matrix
}
\examples{
SPD_matrix(sample(5))
}
