% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SPDts_to_dists}
\alias{SPDts_to_dists}
\title{SPDts_to_dists}
\usage{
SPDts_to_dists(SPDts, states)
}
\arguments{
\item{SPDts}{list of SPD matrices of size n}

\item{states}{list of reference matrices of size d}
}
\value{
A matrix of distances between SPD matrices of size d x n
}
\description{
Computing all the distance between a list of SPD matrices of size n (obtained by the raw data) and a list of reference matrices of size d.
}
