% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{transition_matrix}
\alias{transition_matrix}
\title{transition_matrix}
\usage{
transition_matrix(d = 5, degree = round(d/2), type = "cyclic")
}
\arguments{
\item{d}{dimension of the transition matrix d x d}

\item{degree}{number of connections from one state if \code{type} is \code{"degree"}}

\item{type}{type of matrix to generate: \code{"cyclic"}, \code{"all"}, \code{"degree"}}
}
\value{
A square transition matrix of size d by d. A one value in position (i,j) encodes a transition from state i to state j.
}
\description{
Generating a transition matrix for the change point problem constraint by this transition matrix
}
\examples{
transition_matrix()
}
