% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ts_to_SPDts}
\alias{ts_to_SPDts}
\title{ts_to_SPDts}
\usage{
ts_to_SPDts(data, window_length)
}
\arguments{
\item{data}{matrix of data, size p x n for p dimensions and n data points}

\item{window_length}{the number of data points to take for computing each covariance matrix}
}
\value{
A list of SPD matrices
}
\description{
Converting a time series into a sequence of SPD matrices (covariance matrices)
}
